+++ lib/ui/components/dashboard_card.dart
@@ -10,6 +10,7 @@ import 'package:common/util/extensions/async_ext.dart';
 import 'package:common/util/misc.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:mobileraker/ui/components/console/console_card.dart';
 import 'package:mobileraker/ui/components/power_api_card.dart';
 import 'package:mobileraker_pro/gcode_preview/ui/gcode_preview_card.dart';
 import 'package:mobileraker_pro/service/ui/dashboard_layout_service.dart';
@@ -106,6 +107,7 @@ class DasboardCard extends StatelessWidget {
       DashboardComponentType.firmwareRetraction => FirmwareRetractionCard(machineUUID: machineUUID),
       DashboardComponentType.bedMesh => BedMeshCard(machineUUID: machineUUID),
       DashboardComponentType.gcodePreview => GCodePreviewCard(machineUUID: machineUUID),
+      DashboardComponentType.gcodeConsole => ConsoleCard(machineUUID: machineUUID),
       _ => ErrorCard(
           title: const Text('Unknown card type'),
           body: Text('The card type $type is not supported'),
@@ -133,6 +135,7 @@ class DasboardCard extends StatelessWidget {
       DashboardComponentType.bedMesh => BedMeshCard.preview(),
       DashboardComponentType.webcam => WebcamCard.preview(),
       DashboardComponentType.gcodePreview => GCodePreviewCard.preview(),
+      DashboardComponentType.gcodeConsole => ConsoleCard.preview(),
       _ => Card(
           child: ListTile(
             title: Text(beautifyName(type.name)),